refactor(frontmatter): rename builder to serializer for accuracy
- FrontmatterBuilder → FrontmatterSerializer (reflects actual responsibility)
- build_output() class method → serialize() (clearer intent)
- builder.py → serializer.py (align module name with class name)
- test_builder.py → test_serializer.py (align test file name)
- Keep backward-compatible build_output() function wrapper for external usage

Improves naming clarity: FrontmatterParser parses YAML, FrontmatterSchema
validates, FrontmatterSerializer renders to YAML output. All 12 tests green.
