+++ src/vstack/frontmatter/__init__.py
@@ -1,8 +1,9 @@
 """vstack.frontmatter — YAML frontmatter parsing, building, and schema validation."""
 
-from vstack.frontmatter.builder import build_output as build_output
 from vstack.frontmatter.parser import FrontmatterContent as FrontmatterContent
 from vstack.frontmatter.parser import FrontmatterParser as FrontmatterParser
 from vstack.frontmatter.schema import FieldSpec as FieldSpec
 from vstack.frontmatter.schema import FieldType as FieldType
 from vstack.frontmatter.schema import FrontmatterSchema as FrontmatterSchema
+from vstack.frontmatter.serializer import FrontmatterSerializer as FrontmatterSerializer
+from vstack.frontmatter.serializer import build_output as build_output
